Move GolangRegexValidator to main conftest.py
